make note of what is still broken
the issue is that some footnotes are so long they end up taking multiple
pages - part of the footer is for the long footnote and the rest is for
new footnotes referenced on each new page.

programatically, it is not possible to figure out whether carryover text
from one footnote is part of a long multi-page note or simply from the
last footnote on the previous page, especially given that the way
footnotes are handled from chapter to chapter are not consistent.

these will likely have to be corrected by hand, so for now just note
what pages have broken footnotes.